fix: 채팅 스타일 수정
