+++ client/src/components/chat/chat.module.scss
@@ -2,12 +2,15 @@
 
 .chat_container {
     padding: 17px;
+    width: 250px;
+    box-sizing: content-box;
     border: 1px solid $border-color;
     border-radius: 20px;
     background-color: $base-color;
     display: flex;
     flex-flow: column;
-    box-shadow: 0px 4px 15px 0px rgba(0, 0, 0, 0.15);
+    box-shadow: $normal-shadow;
+    font-family: "Noto Sans KR";
 }
 
 .chat_header {
@@ -16,8 +19,10 @@
     padding-bottom: 15px;
 
     span.chat_text {
-        font-weight: 700;
         font-size: 18px;
+        font-style: normal;
+        font-weight: 700;
+        line-height: normal;
     }
 
     button.close_button {
@@ -34,7 +39,7 @@
     border-bottom: 1px solid $border-color;
     display: flex;
     flex-flow: column nowrap;
-    row-gap: 20px;
+    row-gap: 10px;
     flex-grow: 1;
     overflow-y: auto;
     overflow-x: hidden;
