+++ client/src/components/chat/receivedMessage/receivedMessage.tsx
@@ -2,19 +2,22 @@ import { formatTimeFromISO } from "@/utils";
 import style from "./receivedMessage.module.scss";
 
 interface IMessageProps {
+    nickname: string;
     message: string;
     time: string;
 }
 
 const ReceivedMessage = (props: IMessageProps) => {
-    const { message, time } = props;
+    const { nickname, message, time } = props;
 
     return (
         <div className={style.container}>
+            <div className={style.avatar} />
             <div className={style.wrapper}>
+                <span className={style.nickname}>{nickname}</span>
                 <div className={style.message_wrapper}>
-                    <span className={style.time}>{formatTimeFromISO(time)}</span>
                     <div className={style.message_main}>{message}</div>
+                    <span className={style.time}>{formatTimeFromISO(time)}</span>
                 </div>
             </div>
         </div>
