+++ client/src/styles/global.scss
@@ -1,3 +1,5 @@
+@import "../styles/variables";
+
 body {
     height: 100%;
     overflow-y: hidden;
@@ -21,7 +23,7 @@ button {
 
 /* 스크롤바의 썸네일 스타일 */
 ::-webkit-scrollbar-thumb {
-    background-color: lightgray;
+    background-color: $gray;
     border-radius: 20px;
     border: 2px solid transparent;
     background-clip: padding-box;
@@ -29,5 +31,5 @@ button {
 
 /* 스크롤바 썸네일이 호버 상태일 때 스타일 */
 ::-webkit-scrollbar-thumb:hover {
-    background-color: #b0b0b0;
+    background-color: $gray-dark;
 }
